requirements.txt
Django==2.0
django-authtools==1.6.0
pytz==2017.3
django-countries==5.0
Pillow==4.3.0
django-extensions==1.9.8
django-taggit==0.22.1
silly==1.1.0
ipython
